+++ R/get_elev_raster.R
@@ -64,6 +64,11 @@
 get_elev_raster <- function(locations, z, prj = NULL,src = c("mapzen", "aws"),
                            api_key = get_api_key(src), expand = NULL, ...){
   src <- match.arg(src)
+  if(src=="mapzen"){
+    warning("src 'mapzen' is deprecated and will cease to function after \
+            2018-01-31 due to shutdown of Mapzen; Use 'aws' instead.", 
+            call. = FALSE)
+  }
   # Check location type and if sp, set prj.  If no prj (for either) then error
   locations <- loc_check(locations,prj)
   prj <- sp::proj4string(locations)